pc fixes

+++ imaging/vtkImageMathematics.cxx
@@ -102,25 +102,25 @@ static void vtkImageMathematicsExecute1(vtkImageMathematics *self,
 	    *outPtr = (T)(1.0 / *in1Ptr);
 	    break;
 	  case VTK_SIN:
-	    *outPtr = (T)(sin(*in1Ptr));
+	    *outPtr = (T)(sin((double)*in1Ptr));
 	    break;
 	  case VTK_COS:
-	    *outPtr = (T)(cos(*in1Ptr));
+	    *outPtr = (T)(cos((double)*in1Ptr));
 	    break;
 	  case VTK_EXP:
-	    *outPtr = (T)(exp(*in1Ptr));
+	    *outPtr = (T)(exp((double)*in1Ptr));
 	    break;
 	  case VTK_LOG:
-	    *outPtr = (T)(log(*in1Ptr));
+	    *outPtr = (T)(log((double)*in1Ptr));
 	    break;
 	  case VTK_ABS:
-	    *outPtr = (T)(fabs(*in1Ptr));
+	    *outPtr = (T)(fabs((double)*in1Ptr));
 	    break;
 	  case VTK_SQR:
 	    *outPtr = *in1Ptr * *in1Ptr;
 	    break;
 	  case VTK_SQRT:
-	    *outPtr = (T)(sqrt(*in1Ptr));
+	    *outPtr = (T)(sqrt((double)*in1Ptr));
 	    break;
 	  }
 	outPtr++;
